+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - Advancements (#51)
 - Added a button to ability use messages to apply effects from the ability. (#214)
+- Added a new Active Effect subtype for ability bonuses. (#713)
 - Added a Wealth field to Careers. (#535)
 - Adding or removing a career will adjust the amount of renown and wealth your character has.
 - Added a Renown and Wealth field to characters. (#279)